src/__core/pyproject.toml
[project]
name = "core"
version = "0.0.1"
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "autopep8>=1.6.0",
    "pylint>=2.12.2",
    "flake8>=4.0.1",
    "pydantic>=1.9.0",
    "pytest>=7.0.1",
    "Ipython>=8.1.0",
    "testpath>=0.6.0",
    "djangorestframework>=3.13.1",
    "pytest-cov>=3.0.0",
    "pytest-django>=4.5.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[project.urls]
homepage = ""

[tool]
[tool.pdm]

[tool.pdm.scripts]
pep8 = "autopep8 --in-place --recursive ./src"
pylint = "pylint --disable=C0111 ./src"
test = "pytest --ignore __pypackages__"
test_cov = "pdm run test --cov-fail-under 80 --cov ./src"
test_cov_html = "pytest --ignore __pypackages__ --cov ./src --cov-report html:./__coverage"
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
